requirements.txt
python-telegram-bot==22.2
httpx
python-dotenv
Pillow